backend/requirements.txt
alembic==1.10.2
Flask==2.2.3
google-cloud-pubsub==2.15.2
google-cloud-storage==2.8.0
psycopg2==2.9.6
SQLAlchemy==2.0.8
